requirements.txt
numpy>=1.23
scipy>=1.10
sentence-transformers>=2.2.2
torch>=2.0
faiss-cpu>=1.7.4
pdfplumber>=0.10.3
requests>=2.31.0
tqdm>=4.66.1